Rename match widget menu component to SettingsMenu #297
