+++ app/modules/widgets/__plugins__/match/components/index.js
@@ -2,5 +2,5 @@ export { default } from './__match__'
 export { default as AddChoiceForm } from './add-choice-form'
 export { default as ChoiceCombined } from './choice-combined'
 export { default as Choices } from './choices'
-export { default as Menu } from './menu'
+export { default as SettingsMenu } from './settings-menu'
 export { default as Page } from './page'
