+++ app/modules/widgets/__plugins__/match/components/settings-menu.spec.js
@@ -3,9 +3,9 @@ import { shallow } from 'enzyme'
 import { expect } from 'chai'
 
 import * as Paths from '../../../../../scripts/Paths'
-import Menu from './menu'
+import SettingsMenu from './settings-menu'
 
-describe('app/modules/widgets/__plugins__/match/components/menu', () => {
+describe('app/modules/widgets/__plugins__/match/components/settings-menu', () => {
   let wrapper
 
   const props = {
@@ -18,7 +18,7 @@ describe('app/modules/widgets/__plugins__/match/components/menu', () => {
   const GOALS_TAB = 1
 
   beforeEach(() => {
-    wrapper = shallow(<Menu {...props} />)
+    wrapper = shallow(<SettingsMenu {...props} />)
   })
 
   describe('#render', () => {
